refactor(soft-delete): move BaseRestoreCommand to commands/restore.py
Move the soft-delete restore base class out of commands/base.py
(which is generic command infrastructure) into a sibling
commands/restore.py module. No behaviour change.

Also tighten the TypeVar bound from unbounded T to T bound=SoftDeleteMixin,
which makes the contract — that the base class only restores SoftDeleteMixin
models — visible to mypy and fixes a pre-existing attr-defined error on
model.restore().
