+++ superset/commands/base.py
@@ -15,15 +15,11 @@
 # specific language governing permissions and limitations
 # under the License.
 from abc import ABC, abstractmethod
-from typing import Any, ClassVar, Generic, Optional, TypeVar
+from typing import Any, Optional
 
 from flask_appbuilder.security.sqla.models import User
 
-from superset import security_manager
 from superset.commands.utils import compute_owner_list, populate_owner_list
-from superset.exceptions import SupersetSecurityException
-
-T = TypeVar("T")
 
 
 class BaseCommand(ABC):
@@ -88,59 +84,3 @@ class UpdateMixin:
         :returns: Final list of owners
         """
         return compute_owner_list(current_owners, new_owners)
-
-
-class BaseRestoreCommand(BaseCommand, Generic[T]):
-    """Base class for soft-delete restore commands.
-
-    Subclasses provide the entity-specific bindings as class variables:
-
-      - ``dao``: the DAO class (e.g. ``ChartDAO``)
-      - ``not_found_exc``: exception type raised when the row doesn't
-        exist OR isn't soft-deleted
-      - ``forbidden_exc``: exception type raised when the caller doesn't
-        have ownership of the row
-
-    Subclasses MUST override ``run()`` only to apply the
-    ``@transaction(on_error=partial(on_error, reraise=<RestoreFailed>))``
-    decorator with their concrete restore-failed exception type — the
-    body should be a single ``super().run()`` call. The base class does
-    not enforce this in code; Python decorators don't compose well with
-    class-var-driven configuration. The contract lives in this
-    docstring and in code review.
-
-    The model returned from ``validate()`` is the soft-deleted row,
-    type-narrowed via ``Generic[T]``. ``run()`` calls ``model.restore()``
-    on it (the method comes from ``SoftDeleteMixin``).
-    """
-
-    dao: ClassVar[Any]
-    not_found_exc: ClassVar[type[Exception]]
-    forbidden_exc: ClassVar[type[Exception]]
-
-    def __init__(self, model_uuid: str) -> None:
-        self._model_uuid = model_uuid
-
-    def run(self) -> None:
-        model = self.validate()
-        model.restore()
-
-    def validate(self) -> T:  # type: ignore[override]
-        model = self.dao.find_by_id(
-            self._model_uuid,
-            id_column="uuid",
-            skip_base_filter=True,
-            skip_visibility_filter=True,
-        )
-        if model is None:
-            raise self.not_found_exc(f"No row with uuid={self._model_uuid!r}")
-        if model.deleted_at is None:
-            raise self.not_found_exc(
-                f"Row with uuid={self._model_uuid!r} is not soft-deleted; "
-                f"nothing to restore"
-            )
-        try:
-            security_manager.raise_for_ownership(model)
-        except SupersetSecurityException as ex:
-            raise self.forbidden_exc() from ex
-        return model
